+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/v1/BlueNearBasket.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.auto;
+package org.firstinspires.ftc.teamcode.auto.test.v1;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.Action;
