+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/v1/BlueNearBasketv4.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.auto;
+package org.firstinspires.ftc.teamcode.auto.test.v1;
