+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/v1/CollectTest.java
@@ -1,9 +1,11 @@
-package org.firstinspires.ftc.teamcode.auto;
+package org.firstinspires.ftc.teamcode.auto.test.v1;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.Action;
+import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
+import com.acmerobotics.roadrunner.SleepAction;
 import com.acmerobotics.roadrunner.TrajectoryActionBuilder;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
@@ -14,15 +16,16 @@ import org.firstinspires.ftc.teamcode.MecanumDrive;
 import org.firstinspires.ftc.teamcode.mechanisms.v1.Arm;
 import org.firstinspires.ftc.teamcode.mechanisms.v1.Intake;
 import org.firstinspires.ftc.teamcode.mechanisms.v1.Lift;
+import org.firstinspires.ftc.teamcode.mechanisms.v1.Wrist;
 
 
 @Config
-@Autonomous(name = "\uD83D\uDD34 - OtherRedPark", group = "RoadRunner 1.0")
-public class OtherRedPark extends LinearOpMode {
+@Autonomous(name = "collecttest", group = "RoadRunner 1.0")
+public class CollectTest extends LinearOpMode {
 
 
     // Start position red near
-    Pose2d RED_SCORE_START_POSE = new Pose2d(-36, -60, Math.toRadians(180));
+    Pose2d RED_SCORE_START_POSE = new Pose2d(0, 0, Math.toRadians(0));
 
 
     @Override
@@ -32,35 +35,44 @@ public class OtherRedPark extends LinearOpMode {
         Intake intake = new Intake(hardwareMap);
         Arm arm = new Arm(hardwareMap);
         Lift lift = new Lift(hardwareMap);
+        Wrist wrist = new Wrist(hardwareMap);
         TrajectoryActionBuilder traj = drive.actionBuilder(RED_SCORE_START_POSE)
-                .strafeToLinearHeading(new Vector2d(-55, -55), Math.toRadians(180+45))
-                .strafeToLinearHeading(new Vector2d(-36, -10), Math.toRadians(0))
-                .strafeToLinearHeading(new Vector2d(-26, -10), Math.toRadians(0));
+                .strafeToLinearHeading(new Vector2d(4, 0), Math.toRadians(0));
 
-        Action trajectoryActionCloseOut = traj.fresh()
-                .strafeToLinearHeading(new Vector2d(-36, -12), Math.toRadians(180))
-                .build();
+
+        Action intakeAction = new ParallelAction(
+                traj.build(),
+                intake.intakeAction()
+        );
 
         while(!isStopRequested() && !opModeIsActive()) {
-
         }
 
         waitForStart();
 
         if (isStopRequested()) return;
 
+
         Actions.runBlocking(
                 new SequentialAction(
-                        traj.build(),
-                        arm.armScoreAction(),
-                        lift.liftUpAction(),
-                        intake.depositAction(),
-                        lift.liftDownAction(),
-                        trajectoryActionCloseOut
+                        wrist.wristFoldOutAction(),
+                        arm.armGroundCollectAction(),
+                        intakeAction,
+                        new SleepAction(5),
+                        arm.armRobotTravelAction(),
+                        new SleepAction(2),
+                        arm.armfoldbackaction()
                 )
         );
+
+
+
+
+
+
     } // runOpMode
 
 
 
 }
+
