+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/v1/OtherRedPark.java
@@ -1,11 +1,9 @@
-package org.firstinspires.ftc.teamcode.auto.test;
+package org.firstinspires.ftc.teamcode.auto.test.v1;
 
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.Action;
-import com.acmerobotics.roadrunner.ParallelAction;
 import com.acmerobotics.roadrunner.Pose2d;
 import com.acmerobotics.roadrunner.SequentialAction;
-import com.acmerobotics.roadrunner.SleepAction;
 import com.acmerobotics.roadrunner.TrajectoryActionBuilder;
 import com.acmerobotics.roadrunner.Vector2d;
 import com.acmerobotics.roadrunner.ftc.Actions;
@@ -16,16 +14,15 @@ import org.firstinspires.ftc.teamcode.MecanumDrive;
 import org.firstinspires.ftc.teamcode.mechanisms.v1.Arm;
 import org.firstinspires.ftc.teamcode.mechanisms.v1.Intake;
 import org.firstinspires.ftc.teamcode.mechanisms.v1.Lift;
-import org.firstinspires.ftc.teamcode.mechanisms.v1.Wrist;
 
 
 @Config
-@Autonomous(name = "collecttest", group = "RoadRunner 1.0")
-public class CollectTest extends LinearOpMode {
+@Autonomous(name = "\uD83D\uDD34 - OtherRedPark", group = "RoadRunner 1.0")
+public class OtherRedPark extends LinearOpMode {
 
 
     // Start position red near
-    Pose2d RED_SCORE_START_POSE = new Pose2d(0, 0, Math.toRadians(0));
+    Pose2d RED_SCORE_START_POSE = new Pose2d(-36, -60, Math.toRadians(180));
 
 
     @Override
@@ -35,44 +32,35 @@ public class CollectTest extends LinearOpMode {
         Intake intake = new Intake(hardwareMap);
         Arm arm = new Arm(hardwareMap);
         Lift lift = new Lift(hardwareMap);
-        Wrist wrist = new Wrist(hardwareMap);
         TrajectoryActionBuilder traj = drive.actionBuilder(RED_SCORE_START_POSE)
-                .strafeToLinearHeading(new Vector2d(4, 0), Math.toRadians(0));
+                .strafeToLinearHeading(new Vector2d(-55, -55), Math.toRadians(180+45))
+                .strafeToLinearHeading(new Vector2d(-36, -10), Math.toRadians(0))
+                .strafeToLinearHeading(new Vector2d(-26, -10), Math.toRadians(0));
 
-
-        Action intakeAction = new ParallelAction(
-                traj.build(),
-                intake.intakeAction()
-        );
+        Action trajectoryActionCloseOut = traj.fresh()
+                .strafeToLinearHeading(new Vector2d(-36, -12), Math.toRadians(180))
+                .build();
 
         while(!isStopRequested() && !opModeIsActive()) {
+
         }
 
         waitForStart();
 
         if (isStopRequested()) return;
 
-
         Actions.runBlocking(
                 new SequentialAction(
-                        wrist.wristFoldOutAction(),
-                        arm.armGroundCollectAction(),
-                        intakeAction,
-                        new SleepAction(5),
-                        arm.armRobotTravelAction(),
-                        new SleepAction(2),
-                        arm.armfoldbackaction()
+                        traj.build(),
+                        arm.armScoreAction(),
+                        lift.liftUpAction(),
+                        intake.depositAction(),
+                        lift.liftDownAction(),
+                        trajectoryActionCloseOut
                 )
         );
-
-
-
-
-
-
     } // runOpMode
 
 
 
 }
-
