+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/v1/RedNearBasketv3.java
@@ -1,6 +1,4 @@
-package org.firstinspires.ftc.teamcode.auto;
-
-
+package org.firstinspires.ftc.teamcode.auto.test.v1;
 import com.acmerobotics.dashboard.config.Config;
 import com.acmerobotics.roadrunner.Action;
 import com.acmerobotics.roadrunner.ParallelAction;
@@ -21,36 +19,37 @@ import org.firstinspires.ftc.teamcode.mechanisms.v1.Wrist;
 
 
 @Config
-@Autonomous(name = "\uD83D\uDD35 - RedNearBasketV3", group = "RoadRunner 1.0")
-public class BlueNearBasketv3 extends LinearOpMode {
+@Autonomous(name = "\uD83D\uDD34 - Red+BlueNearBasketV3", group = "Meet2", preselectTeleOp = "TeleOpModular")
+public class RedNearBasketv3 extends LinearOpMode {
 
 
     // Start position red near
-    Pose2d BLUE_SCORE_START_POSE = new Pose2d(-38, -60, Math.toRadians(180));
+    Pose2d RED_SCORE_START_POSE = new Pose2d(-38, -60, Math.toRadians(180));
 
-    public static double BLUE_BASKET_X = 47;
-    public static double BLUE_BASKET_Y = 47;
-    public static double BLUE_BASKET_HEADING = 45;
+    public static double RED_BASKET_X = -47;
+    public static double RED_BASKET_Y = -47;
+    public static double RED_BASKET_HEADING = 180+45;
 
-    public static double BLUE_SAMPLE1_X = 26;
-    public static double BLUE_SAMPLE2_X = 35;
-    public static double BLUE_SAMPLE3_X = 42; // -46 would hit the boundary
+    public static double RED_SAMPLE1_X = -26;
+    public static double RED_SAMPLE2_X = -35;
+    public static double RED_SAMPLE3_X = -42; // -46 would hit the boundary
 
-    public static double BLUE_SAMPLE1_Y = 30;
-    public static double BLUE_SAMPLE2_Y = 23;
-    public static double BLUE_SAMPLE3_Y = 23;
+    public static double RED_SAMPLE1_Y = -30;
+    public static double RED_SAMPLE2_Y = -23;
+    public static double RED_SAMPLE3_Y = -23;
 
-    public static double BLUE_SAMPLE1_HEADING = -161;
-    public static double RED_SAMPLE2_HEADING = 0;
-    public static double RED_SAMPLE3_HEADING = 0;
+    public static double RED_SAMPLE1_HEADING = 161;
+    public static double RED_SAMPLE2_HEADING = 180;
+    public static double RED_SAMPLE3_HEADING = 180;
 
 
     @Override
     public void runOpMode() {
 
-        MecanumDrive drive = new MecanumDrive(hardwareMap, BLUE_SCORE_START_POSE);
+        MecanumDrive drive = new MecanumDrive(hardwareMap, RED_SCORE_START_POSE);
         Intake intake = new Intake(hardwareMap);
         Arm arm = new Arm(hardwareMap);
+        arm.reset();
         Lift lift = new Lift(hardwareMap);
         Wrist wrist = new Wrist(hardwareMap);
 
@@ -58,9 +57,13 @@ public class BlueNearBasketv3 extends LinearOpMode {
         // TODO: Move to a shared file
         Action scoreHighAction = new SequentialAction(
                 new ParallelAction(
-                        wrist.wristFoldOutAction(),
-                        arm.armScoreAction()),
+                    wrist.wristFoldOutAction(),
+                    arm.armScoreAction()),
                 lift.liftUpAction()
+        );
+        Action armgopark = new ParallelAction(
+                lift.liftDownAction(),
+                arm.armParkAction()
         );
 //        Action scoreHighAction2 = new ParallelAction(
 //                arm.armScoreAction(),
@@ -90,34 +93,38 @@ public class BlueNearBasketv3 extends LinearOpMode {
         // ==== End of Non-trajectory related actions ====
 
         // ==== Start of Trajectory actions ====
-        TrajectoryActionBuilder startToBasketTab = drive.actionBuilder(BLUE_SCORE_START_POSE)
+        TrajectoryActionBuilder startToBasketTab = drive.actionBuilder(RED_SCORE_START_POSE)
                 .strafeToLinearHeading(new Vector2d(-38, -56), Math.toRadians(180))
-                .strafeToLinearHeading(new Vector2d(BLUE_BASKET_X, BLUE_BASKET_Y), Math.toRadians(BLUE_BASKET_HEADING));
+                .strafeToLinearHeading(new Vector2d(RED_BASKET_X, RED_BASKET_Y), Math.toRadians(RED_BASKET_HEADING));
 
         TrajectoryActionBuilder driveBasketToSample1Tab = startToBasketTab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_SAMPLE1_X, BLUE_SAMPLE1_Y), Math.toRadians(BLUE_SAMPLE1_HEADING));
+                .strafeToLinearHeading(new Vector2d(RED_SAMPLE1_X, RED_SAMPLE1_Y), Math.toRadians(RED_SAMPLE1_HEADING));
 
         TrajectoryActionBuilder pickUpSample1Tab = driveBasketToSample1Tab.endTrajectory().fresh()
-                .setTangent(Math.toRadians(BLUE_SAMPLE1_HEADING))
-                .lineToX(BLUE_SAMPLE1_X - 4);
+                .setTangent(Math.toRadians(RED_SAMPLE1_HEADING))
+                .lineToX(RED_SAMPLE1_X - 4);
 
         TrajectoryActionBuilder driveSample1ToBasketTab = pickUpSample1Tab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_BASKET_X, BLUE_BASKET_Y), Math.toRadians(BLUE_BASKET_HEADING));
+                .strafeToLinearHeading(new Vector2d(RED_BASKET_X, RED_BASKET_Y), Math.toRadians(RED_BASKET_HEADING));
 
         TrajectoryActionBuilder driveBasketToSample2Tab = driveSample1ToBasketTab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_SAMPLE2_X, BLUE_SAMPLE2_Y), Math.toRadians(180));
+                .strafeToLinearHeading(new Vector2d(RED_SAMPLE2_X, RED_SAMPLE2_Y), Math.toRadians(180));
 
         TrajectoryActionBuilder pickUpSample2Tab = driveBasketToSample2Tab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_SAMPLE2_X - 4, BLUE_SAMPLE2_Y), Math.toRadians(180));
+                .strafeToLinearHeading(new Vector2d(RED_SAMPLE2_X - 4, RED_SAMPLE2_Y), Math.toRadians(180));
 
         TrajectoryActionBuilder driveSample2ToBasketTab = pickUpSample2Tab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_BASKET_X, BLUE_BASKET_Y), Math.toRadians(BLUE_BASKET_HEADING));
+                .strafeToLinearHeading(new Vector2d(RED_BASKET_X, RED_BASKET_Y), Math.toRadians(RED_BASKET_HEADING));
 
         TrajectoryActionBuilder driveBasketToSample3Tab = driveSample2ToBasketTab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_SAMPLE3_X, BLUE_SAMPLE3_Y), Math.toRadians(180));
+                .strafeToLinearHeading(new Vector2d(RED_SAMPLE3_X, RED_SAMPLE3_Y), Math.toRadians(180));
 
         TrajectoryActionBuilder driveSample3ToBasketTab = driveBasketToSample3Tab.endTrajectory().fresh()
-                .strafeToLinearHeading(new Vector2d(BLUE_BASKET_X, BLUE_BASKET_Y), Math.toRadians(BLUE_BASKET_HEADING));
+                .strafeToLinearHeading(new Vector2d(RED_BASKET_X, RED_BASKET_Y), Math.toRadians(RED_BASKET_HEADING));
+
+        TrajectoryActionBuilder driveBasketToPark = driveSample3ToBasketTab.endTrajectory().fresh()
+                .strafeToLinearHeading(new Vector2d(-33,-9),Math.toRadians(180))
+                .strafeToLinearHeading(new Vector2d(-22,-9),Math.toRadians(180));
 
         // ==== End of Trajectory actions ====
 
@@ -158,8 +165,8 @@ public class BlueNearBasketv3 extends LinearOpMode {
 
         Action cSample1ToBasketAction = new SequentialAction(
                 new ParallelAction(
-                        driveSample1ToBasketTab.build(),
-                        arm.armScoreAction()),
+                    driveSample1ToBasketTab.build(),
+                    arm.armScoreAction()),
                 lift.liftUpAction()
         );
         Action cSample2ToBasketAction = new SequentialAction(
@@ -184,6 +191,14 @@ public class BlueNearBasketv3 extends LinearOpMode {
                 pickUpSample2Tab.build(),
                 arm.armGroundCollectAction(),
                 intake.intakeAction()
+        );
+
+        Action cBasketToPark = new ParallelAction(
+                new SequentialAction(
+                        arm.armVerticalAction()
+                        ,lift.liftDownAction()
+                ),
+                driveBasketToPark.build()
         );
 
 
@@ -217,6 +232,8 @@ public class BlueNearBasketv3 extends LinearOpMode {
                         cSample2ToBasketAction,
                         new SleepAction(0.1),
                         intake.depositAction(),
+                        cBasketToPark,
+                        arm.armParkAction()
                         //
 //                        cBasketToSample3Action,
 //                        arm.armGroundCollectAction(),
@@ -226,7 +243,7 @@ public class BlueNearBasketv3 extends LinearOpMode {
 //                        new SleepAction(0.1),
 //                        intake.depositAction(),
                         //
-                        foldBackAction
+//                        foldBackAction
                 )
         );
 
@@ -235,5 +252,4 @@ public class BlueNearBasketv3 extends LinearOpMode {
 
 
 }
-
