+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/auto/test/v1/WristTest.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.auto.test;
+package org.firstinspires.ftc.teamcode.auto.test.v1;
 
 
 import com.acmerobotics.dashboard.config.Config;
